Can pass duration as a query parameter
Added the ability to pass duration as a query parameter. You can called index.html?d=60 if you want the timer to be set for 60 seconds. Defaults to 5 mins if nothing is passed.

Also added a 200px padding on top to keep the clock in the centre on a laptop screen. (Guess this might cause a problem on phone.)